+++ host/adblock/RELEASE_NOTES.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 - Removed "Always allow ads on this site" from kebab menu (EXT-595)
+- Added skeleton for public API inside custom iframe; the API is currently inactive (EXT-567)
 
 # 6.12.0 - 2024-11-25